refactor: move ansible related code to its own files
