+++ src/v1/mocks/Ansible.js
@@ -1,0 +1,166 @@
+/* Copyright (c) 2021 Red Hat, Inc. */
+/* Copyright Contributors to the Open Cluster Management project */
+
+export const mockAnsibleSecretsResponse = {
+
+};
+
+export const mockAnsibleJobTemplatesResponse = {
+  count: 1,
+  next: null,
+  previous: null,
+  results: [
+    {
+      id: 7,
+      type: 'job_template',
+      url: '/api/v2/job_templates/7/',
+      related: {
+        created_by: '/api/v2/users/1/',
+        modified_by: '/api/v2/users/1/',
+        labels: '/api/v2/job_templates/7/labels/',
+        inventory: '/api/v2/inventories/1/',
+        project: '/api/v2/projects/6/',
+        organization: '/api/v2/organizations/1/',
+        credentials: '/api/v2/job_templates/7/credentials/',
+        jobs: '/api/v2/job_templates/7/jobs/',
+        schedules: '/api/v2/job_templates/7/schedules/',
+        activity_stream: '/api/v2/job_templates/7/activity_stream/',
+        launch: '/api/v2/job_templates/7/launch/',
+        webhook_key: '/api/v2/job_templates/7/webhook_key/',
+        webhook_receiver: '',
+        notification_templates_started: '/api/v2/job_templates/7/notification_templates_started/',
+        notification_templates_success: '/api/v2/job_templates/7/notification_templates_success/',
+        notification_templates_error: '/api/v2/job_templates/7/notification_templates_error/',
+        access_list: '/api/v2/job_templates/7/access_list/',
+        survey_spec: '/api/v2/job_templates/7/survey_spec/',
+        object_roles: '/api/v2/job_templates/7/object_roles/',
+        instance_groups: '/api/v2/job_templates/7/instance_groups/',
+        slice_workflow_jobs: '/api/v2/job_templates/7/slice_workflow_jobs/',
+        copy: '/api/v2/job_templates/7/copy/',
+      },
+      summary_fields: {
+        organization: {
+          id: 1,
+          name: 'Default',
+          description: '',
+        },
+        inventory: {
+          id: 1,
+          name: 'Demo Inventory',
+          description: '',
+          has_active_failures: false,
+          total_hosts: 0,
+          hosts_with_active_failures: 0,
+          total_groups: 0,
+          has_inventory_sources: false,
+          total_inventory_sources: 0,
+          inventory_sources_with_failures: 0,
+          organization_id: 1,
+          kind: '',
+        },
+        project: {
+          id: 6,
+          name: 'Demo Project',
+          description: '',
+          status: 'never updated',
+          scm_type: 'git',
+        },
+        created_by: {
+          id: 1,
+          username: 'admin',
+          first_name: '',
+          last_name: '',
+        },
+        modified_by: {
+          id: 1,
+          username: 'admin',
+          first_name: '',
+          last_name: '',
+        },
+        object_roles: {
+          admin_role: {
+            description: 'Can manage all aspects of the job template',
+            name: 'Admin',
+            id: 30,
+          },
+          execute_role: {
+            description: 'May run the job template',
+            name: 'Execute',
+            id: 31,
+          },
+          read_role: {
+            description: 'May view settings for the job template',
+            name: 'Read',
+            id: 32,
+          },
+        },
+        user_capabilities: {
+          edit: true,
+          delete: true,
+          start: true,
+          schedule: true,
+          copy: true,
+        },
+        labels: {
+          count: 0,
+          results: [
+          ],
+        },
+        recent_jobs: [
+        ],
+        credentials: [
+          {
+            id: 1,
+            name: 'Demo Credential',
+            description: '',
+            kind: 'ssh',
+            cloud: false,
+          },
+        ],
+      },
+      created: '2021-04-27T22:10:35.914847Z',
+      modified: '2021-04-27T22:10:35.914878Z',
+      name: 'Demo Job Template',
+      description: '',
+      job_type: 'run',
+      inventory: 1,
+      project: 6,
+      playbook: 'hello_world.yml',
+      scm_branch: '',
+      forks: 0,
+      limit: '',
+      verbosity: 0,
+      extra_vars: '',
+      job_tags: '',
+      force_handlers: false,
+      skip_tags: '',
+      start_at_task: '',
+      timeout: 0,
+      use_fact_cache: false,
+      organization: 1,
+      last_job_run: null,
+      last_job_failed: false,
+      next_job_run: null,
+      status: 'never updated',
+      host_config_key: '',
+      ask_scm_branch_on_launch: false,
+      ask_diff_mode_on_launch: false,
+      ask_variables_on_launch: false,
+      ask_limit_on_launch: false,
+      ask_tags_on_launch: false,
+      ask_skip_tags_on_launch: false,
+      ask_job_type_on_launch: false,
+      ask_verbosity_on_launch: false,
+      ask_inventory_on_launch: false,
+      ask_credential_on_launch: false,
+      survey_enabled: false,
+      become_enabled: false,
+      diff_mode: false,
+      allow_simultaneous: false,
+      custom_virtualenv: null,
+      job_slice_count: 1,
+      webhook_service: '',
+      webhook_credential: null,
+    },
+  ],
+};
