+++ src/v1/schema/ansible.test.js
@@ -1,0 +1,32 @@
+/* Copyright (c) 2021 Red Hat, Inc. */
+/* Copyright Contributors to the Open Cluster Management project */
+
+import supertest from 'supertest';
+import nock from 'nock';
+import server, { GRAPHQL_PATH } from '../index';
+import {
+  mockAnsibleJobTemplatesResponse,
+} from '../mocks/Ansible';
+
+describe('Ansible Automation Resolver', () => {
+  test('Correctly Resolves Ansible Job Templates', () => new Promise((done) => {
+    nock('https://ansible-tower.com').persist().get('/api/v2/job_templates')
+      .reply(200, mockAnsibleJobTemplatesResponse);
+    supertest(server)
+      .post(GRAPHQL_PATH)
+      .send({
+        query: `{
+          ansibleJobTemplates(towerURL:"https://ansible-tower.com", token:"mocktoken"){
+            name
+            description
+            extra_vars
+          }
+        }
+      `,
+      })
+      .end((err, res) => {
+        expect(JSON.parse(res.text)).toMatchSnapshot();
+        done();
+      });
+  }));
+});
